lf chat: make follow mode the sole human thread surface
